rework water requests, create utils for unixDate, add rendering after requests, add new path to backend
