+++ src/components/DailyStats/DailyStats.jsx
@@ -1,42 +1,23 @@
-import React, { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
+
+import { useState } from 'react';
+import { selectActiveDay, selectDayWater } from '../../redux/water/selectors';
 import css from './DailyStats.module.css';
 import WaterList from '../WaterList/WaterList';
 import AddWaterBtn from '../AddWaterBtn/AddWaterBtn';
 import Modal from '../../shared/Modal/Modal';
 import EditWaterModal from '../Modals/EditWaterModal/EditWaterModal';
 import DeleteModal from '../Modals/DeleteModal/DeleteModal';
-import {
-  getDayWater,
-  deleteWater,
-  updateWater,
-} from '../../redux/water/operations';
-import { selectActiveDay, selectDayWater } from '../../redux/water/selectors';
-import { setActiveDay } from '../../redux/water/slice';
+import { deleteWater, updateWater } from '../../redux/water/operations';
 
 export default function DailyStats() {
   const dispatch = useDispatch();
-
   const activeDay = useSelector(selectActiveDay);
   const dayWater = useSelector(selectDayWater)?.WaterData || [];
 
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [currentItem, setCurrentItem] = useState(null);
-
-  useEffect(() => {
-    if (!activeDay) {
-      const now = new Date();
-      const currentDate = now.toISOString();
-      dispatch(setActiveDay(currentDate));
-    }
-  }, [dispatch, activeDay]);
-
-  useEffect(() => {
-    if (activeDay) {
-      dispatch(getDayWater(activeDay));
-    }
-  }, [activeDay, dispatch]);
 
   const openEditModal = item => {
     setCurrentItem(item);
