+++ src/components/WaterProgressBar/WaterProgressBar.jsx
@@ -1,28 +1,12 @@
-import { useSelector, useDispatch } from 'react-redux';
-import { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+import { selectUserWaterNorm } from '../../redux/user/selectors';
+import { selectDayWater } from '../../redux/water/selectors';
 import Icon from '../../shared/Icons/Icon';
 import css from './WaterProgressBar.module.css';
 
-import { selectUserWaterNorm } from '../../redux/user/selectors';
-import { selectDayWater, selectActiveDay } from '../../redux/water/selectors';
-import { getDayWater } from '../../redux/water/operations';
-
 const WaterProgressBar = () => {
-  const dispatch = useDispatch();
-  const activeDay = useSelector(selectActiveDay);
   const dailyWaterNorm = useSelector(selectUserWaterNorm);
   const dayWater = useSelector(selectDayWater);
-
-  useEffect(() => {
-    if (activeDay) {
-      const timestamp = Date.parse(activeDay);
-      if (!isNaN(timestamp)) {
-        dispatch(getDayWater(timestamp));
-      } else {
-        console.error('Invalid date format in activeDay:', activeDay);
-      }
-    }
-  }, [activeDay, dispatch]);
 
   const totalDrinkingWater = dayWater?.totalDayWater || 0;
 
@@ -30,20 +14,11 @@ const WaterProgressBar = () => {
     ? Math.min(Math.round((totalDrinkingWater / dailyWaterNorm) * 100), 100)
     : 0;
 
-  const [indicatorPosition, setIndicatorPosition] = useState(0);
-
-  useEffect(() => {
-    setIndicatorPosition(percentage);
-  }, [percentage]);
-
   return (
     <div className={css.progressBarWrapp}>
       <h4 className={css.progressBarText}>Today</h4>
       <div className={css.progressBar}>
-        <div
-          className={css.progressline}
-          style={{ width: `${indicatorPosition}%` }}
-        >
+        <div className={css.progressline} style={{ width: `${percentage}%` }}>
           <Icon
             className={css.progressIndicator}
             id="elipce"
