+++ src/redux/water/slice.js
@@ -35,7 +35,6 @@ const waterSlice = createSlice({
       })
       .addCase(getDayWater.fulfilled, (state, action) => {
         state.loading = false;
-
         state.dayWater = {
           ...action.payload,
           water: Array.isArray(action.payload.water)
@@ -45,6 +44,7 @@ const waterSlice = createSlice({
       })
       .addCase(getDayWater.rejected, (state, action) => {
         state.loading = false;
+        state.dayWater = { date: '', water: [] };
         state.error = action.payload;
       })
 
@@ -69,11 +69,6 @@ const waterSlice = createSlice({
       })
       .addCase(addWater.fulfilled, (state, action) => {
         state.loading = false;
-
-        if (!Array.isArray(state.dayWater.water)) {
-          state.dayWater.water = [];
-        }
-
         state.dayWater.water.push(action.payload);
       })
       .addCase(addWater.rejected, (state, action) => {
@@ -88,10 +83,6 @@ const waterSlice = createSlice({
       })
       .addCase(updateWater.fulfilled, (state, action) => {
         state.loading = false;
-
-        state.dayWater.water = state.dayWater.water.map(item =>
-          item._id === action.payload._id ? action.payload : item
-        );
       })
       .addCase(updateWater.rejected, (state, action) => {
         state.loading = false;
@@ -105,10 +96,6 @@ const waterSlice = createSlice({
       })
       .addCase(deleteWater.fulfilled, (state, action) => {
         state.loading = false;
-
-        state.dayWater.water = state.dayWater.water.filter(
-          item => item._id !== action.payload._id
-        );
       })
       .addCase(deleteWater.rejected, (state, action) => {
         state.loading = false;
